refactor: separate connection service and host ownership
